Moved the Organism Files into the Entities package and refactored. Also added some needed diagram files.
