+++ doc/UMLdiagrams/OrganismFactory.plantuml
@@ -1,0 +1,16 @@
+@startuml
+
+!theme plain
+top to bottom direction
+skinparam linetype ortho
+
+class OrganismFactory {
+  - OrganismFactory(): 
+  - m_RegisteredOrganisms: HashMap<String, AbstractOrganism>
+  - singleton: OrganismFactory
+  + registerOrganism(String, AbstractOrganism): void
+  + createOrganism(String): AbstractOrganism
+  + getInstance(): OrganismFactory
+}
+
+@enduml
