+++ src/main/java/team/rocket/Entities/Grass.java
@@ -1,22 +1,30 @@
-package team.rocket;
+package team.rocket.Entities;
 
 import team.rocket.Enums.Direction;
-
 import java.util.Random;
-public class Rabbit extends AbstractAnimal{
-    private static final char icon = 'R';
-    private static int count = 0;
-    private boolean hasMoved;
-    private int food;
-    private static int deathFood = 0;
-    private static int foodIncrement = 1;
 
-    public Rabbit(){
+public class Grass extends AbstractPlant {
+    private static char icon = 'G';
+    private static int count;
+    private boolean hasGrown;
+
+    /**
+     * Creates new grass, not ready to grow
+     */
+    public Grass(){
         count++;
+        hasGrown = true;
     }
 
     /**
-     * @return team.rocket.Rabbit's icon as a character
+     * @return boolean indication if grass has grown this cycle
+     */
+    public boolean growthStatus(){
+        return this.hasGrown;
+    }
+
+    /**
+     * @return Grass's icon as a character
      */
     public static char toIcon(){
         return icon;
@@ -29,46 +37,48 @@ public class Rabbit extends AbstractAnimal{
     public char instancedToIcon(){return icon;}
 
     /**
-     * @return current Rabbit count
+     * @return current Grass count
      */
     public static int getCount(){
         return count;
     }
 
-    @Override
+    /**
+     * Sets the count of Grass
+     * @param i the number count is being set too
+     */
     public void setCount(int i) {
         count = i;
     }
 
-
-    @Override
+    /**
+     * Needed for very specific instance with OrganismEnum so that the instance in the enum doesn't count towards the total number of Organisms
+     */
     public void reduceCount() {
         count--;
     }
 
-    @Override
-    public AbstractOrganism getNewObjectFromExistingObject() {
-        return new Rabbit();
+    /**
+     * Takes the instance of an object and creates a brand new one and returns that new object
+     * @return a fresh new not-copied AbstractOrganism
+     */
+    public AbstractOrganism getNewObjectFromExistingObject(){
+        return new Grass();
     }
 
     /**
-     * Creates new Rabbit
+     * Resets hasGrown to false, meant to be used to reset growth each day
      */
-    public void reproduce(){} //not yet implemented
-
-    /**
-     * Resets hasMoved to false, meant to be used to reset movement each day
-     */
-    public void resetMove(){
-        this.hasMoved = false;
+    public void resetGrown(){
+        this.hasGrown = false;
     }
 
     /**
-     * Takes array of a team.rocket.Rabbit's neighbors, randomly chooses an available space, and returns corresponding direction
+     * Takes array of a team.rocket.Entities.Rabbit's neighbors, randomly chooses an available space, and returns corresponding direction
      * @param neighbors array of animals in adjacent tiles, 0-3 representing UP, DOWN, LEFT, or RIGHT respectively
      * @return randomly determined direction based on available spaces
      */
-    public Direction availableMovementSpace(AbstractOrganism[] neighbors){
+    public Direction availableSpace(AbstractOrganism[] neighbors){
         int i = 0; //tracks iterations of for loop
         int freeSpaceCount = 0; //stores number of free adjacent spaces
         Direction[] freeSpaces = new Direction[4]; //stores available movement directions
@@ -108,51 +118,38 @@ public class Rabbit extends AbstractAnimal{
     }
 
     /**
-     * Moves Rabbit in grid based on current position, available movement space, and past movement
+     * Creates new Grass in free adjacent slot
      * @param grid 2D array holding all Organisms in simulation
      * @param neighbors array of animals in adjacent tiles, 0-3 representing UP, DOWN, LEFT, or RIGHT respectively
-     * @param y - y position of Rabbit in grid
-     * @param x - x position of Rabbit in grid
+     * @param y - y position of Grass in grid
+     * @param x - x position of Grass in grid
      */
-    public void move(AbstractAnimal[][] grid, AbstractAnimal[] neighbors, int y, int x) {
-        if (hasMoved) {
+    public void grow(AbstractOrganism grid[][], AbstractOrganism[] neighbors, int y, int x) {
+        if (hasGrown) {
             return;
         }
 
-        Direction direction = this.availableMovementSpace(neighbors);
+        Direction direction = this.availableSpace(neighbors);
 
         if (direction == null) {
             return;
         }
 
         if (direction == Direction.UP) {
-            grid[y][x] = null;
-            grid[y-1][x] = this;
+            grid[y-1][x] = getNewObjectFromExistingObject();
         }
 
         if (direction == Direction.DOWN) {
-            grid[y][x] = null;
-            grid[y+1][x] = this;
+            grid[y+1][x] = getNewObjectFromExistingObject();
         }
 
         if (direction == Direction.LEFT) {
-            grid[y][x] = null;
-            grid[y][x-1] = this;
+            grid[y][x-1] = getNewObjectFromExistingObject();
         }
 
         if (direction == Direction.RIGHT) {
-            grid[y][x] = null;
-            grid[y][x+1] = this;
+            grid[y][x+1] = getNewObjectFromExistingObject();
         }
-        hasMoved = true;
-    }
-
-    public boolean isStarving() {
-        return food < deathFood;
-    }
-
-    public void eat(Map map, int row, int column) {
-        map.removeOrganism(row, column);
-        food += foodIncrement;
+        hasGrown = true;
     }
 }
