+++ src/main/java/team/rocket/Entities/OrganismFactory.java
@@ -1,4 +1,4 @@
-package team.rocket;
+package team.rocket.Entities;
 
 import java.util.HashMap;
