+++ src/test/java/stepdefinitions/Grass.java
@@ -3,18 +3,18 @@ package stepdefinitions;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import org.junit.jupiter.api.Assertions;
-import team.rocket.AbstractOrganism;
+import team.rocket.Entities.AbstractOrganism;
 
 public class Grass {
-    team.rocket.Grass g;
-    team.rocket.Grass g2;
-    team.rocket.AbstractOrganism grid[][];
+    team.rocket.Entities.Grass g;
+    team.rocket.Entities.Grass g2;
+    AbstractOrganism grid[][];
     AbstractOrganism[] neighbors;
     int y;
     int x;
 
     @Given("A new grass is grown")
-    public void aNewGrassIsGrown(){g = new team.rocket.Grass();}
+    public void aNewGrassIsGrown(){g = new team.rocket.Entities.Grass();}
 
     @Then("The grass should be unable to grow")
     public void theGrassShouldBeUnableToGrow(){Assertions.assertEquals(true, g.growthStatus());}
