+++ src/test/java/stepdefinitions/Map.java
@@ -4,7 +4,8 @@ import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import team.rocket.AbstractOrganism;
+import team.rocket.Entities.AbstractOrganism;
+import team.rocket.Entities.Rabbit;
 
 import java.util.Arrays;
 
@@ -43,7 +44,7 @@ public class Map {
 
     @Given("a map is created with a grid")
     public void aMapIsCreatedWithAGrid() {
-        team.rocket.AbstractOrganism[][] grid = new team.rocket.AbstractOrganism[2][2];
+        AbstractOrganism[][] grid = new AbstractOrganism[2][2];
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertNotNull(map);
@@ -51,7 +52,7 @@ public class Map {
 
     @Then("the map should have the given dimensions and given contents")
     public void theMapShouldHaveTheGivenDimensionsAndGivenContents() {
-        team.rocket.AbstractOrganism[][] grid = new team.rocket.AbstractOrganism[2][2];
+        AbstractOrganism[][] grid = new AbstractOrganism[2][2];
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertTrue(Arrays.deepEquals(grid, map.getGrid()));
@@ -63,7 +64,7 @@ public class Map {
 
     @Given("an empty map is created")
     public void anEmptyMapIsCreated() {
-        team.rocket.AbstractOrganism[][] grid = new AbstractOrganism[1][1];
+        AbstractOrganism[][] grid = new AbstractOrganism[1][1];
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertNotNull(map);
@@ -71,7 +72,7 @@ public class Map {
 
     @Then("the map should be empty")
     public void theMapShouldBeEmpty() {
-        team.rocket.AbstractOrganism[][] grid = new AbstractOrganism[1][1];
+        AbstractOrganism[][] grid = new AbstractOrganism[1][1];
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertTrue(map.isEmpty());
@@ -79,7 +80,7 @@ public class Map {
 
     @Given("a map of all rabbits is created")
     public void aMapOfAllRabbitsIsCreated() {
-        team.rocket.AbstractOrganism[][] grid = new team.rocket.AbstractOrganism[][] {{new team.rocket.Rabbit()}};
+        AbstractOrganism[][] grid = new AbstractOrganism[][] {{new Rabbit()}};
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertNotNull(map);
@@ -87,7 +88,7 @@ public class Map {
 
     @Then("the map should be full")
     public void theMapShouldBeFull() {
-        team.rocket.AbstractOrganism[][] grid = new team.rocket.AbstractOrganism[][] {{new team.rocket.Rabbit()}};
+        AbstractOrganism[][] grid = new AbstractOrganism[][] {{new Rabbit()}};
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertTrue(map.isFull());
@@ -95,7 +96,7 @@ public class Map {
 
     @Given("a nonempty map is created")
     public void aNonemptyMapIsCreated() {
-        team.rocket.AbstractOrganism[][] grid = new team.rocket.AbstractOrganism[][] {{null, null}, {new team.rocket.Rabbit(), null}};
+        AbstractOrganism[][] grid = new AbstractOrganism[][] {{null, null}, {new Rabbit(), null}};
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertNotNull(map);
@@ -103,7 +104,7 @@ public class Map {
 
     @Then("the map should not be empty")
     public void theMapShouldNotBeEmpty() {
-        team.rocket.AbstractOrganism[][] grid = new team.rocket.AbstractOrganism[][] {{null, null}, {new team.rocket.Rabbit(), null}};
+        AbstractOrganism[][] grid = new AbstractOrganism[][] {{null, null}, {new Rabbit(), null}};
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertFalse(map.isEmpty());
@@ -111,7 +112,7 @@ public class Map {
 
     @Given("a nonfull map is created")
     public void aNonfullMapIsCreated() {
-        team.rocket.AbstractOrganism[][] grid = new team.rocket.AbstractOrganism[][] {{new team.rocket.Rabbit(), null}, {new team.rocket.Rabbit(), new team.rocket.Rabbit()}};
+        AbstractOrganism[][] grid = new AbstractOrganism[][] {{new Rabbit(), null}, {new Rabbit(), new Rabbit()}};
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertNotNull(map);
@@ -119,7 +120,7 @@ public class Map {
 
     @Then("the map should not be full")
     public void theMapShouldNotBeFull() {
-        team.rocket.AbstractOrganism[][] grid = new team.rocket.AbstractOrganism[][] {{new team.rocket.Rabbit(), null}, {new team.rocket.Rabbit(), new team.rocket.Rabbit()}};
+        AbstractOrganism[][] grid = new AbstractOrganism[][] {{new Rabbit(), null}, {new Rabbit(), new Rabbit()}};
         team.rocket.Map map = new team.rocket.Map(grid);
 
         assertFalse(map.isFull());
@@ -137,14 +138,14 @@ public class Map {
     @When("an organism is added to the space")
     public void anOrganismIsAddedToTheSpace() {
         team.rocket.Map map = new team.rocket.Map(2, 2);
-        team.rocket.AbstractOrganism rabbit = new team.rocket.Rabbit();
+        AbstractOrganism rabbit = new Rabbit();
         map.addOrganism(rabbit, 1, 1);
     }
 
     @Then("the space should have the given organism")
     public void theSpaceShouldHaveTheGivenOrganism() {
         team.rocket.Map map = new team.rocket.Map(2, 2);
-        team.rocket.AbstractOrganism rabbit = new team.rocket.Rabbit();
+        AbstractOrganism rabbit = new Rabbit();
         map.addOrganism(rabbit, 1, 1);
 
         assertEquals(rabbit, map.getGrid()[1][1]);
@@ -153,14 +154,14 @@ public class Map {
     @Given("a space has an organism")
     public void aSpaceHasAnOrganism() {
         team.rocket.Map map = new team.rocket.Map(2, 2);
-        team.rocket.AbstractOrganism rabbit = new team.rocket.Rabbit();
+        AbstractOrganism rabbit = new Rabbit();
         map.addOrganism(rabbit, 1, 1);
     }
 
     @When("an organism is removed from the space")
     public void anOrganismIsRemovedFromTheSpace() {
         team.rocket.Map map = new team.rocket.Map(2, 2);
-        team.rocket.AbstractOrganism rabbit = new team.rocket.Rabbit();
+        AbstractOrganism rabbit = new Rabbit();
         map.addOrganism(rabbit, 1, 1);
         map.removeOrganism(1, 1);
     }
@@ -168,7 +169,7 @@ public class Map {
     @Then("the space should be empty")
     public void theSpaceShouldBeEmpty() {
         team.rocket.Map map = new team.rocket.Map(2, 2);
-        team.rocket.AbstractOrganism rabbit = new team.rocket.Rabbit();
+        AbstractOrganism rabbit = new Rabbit();
         map.addOrganism(rabbit, 1, 1);
         map.removeOrganism(1, 1);
